handle submit email
prevent default form submit functionality and send using javascript. Sends the POST to an AWS API Gateway
on the Stylepoints AWS account which adds the Mailchimp Basic Authentication header and then proxies the request
to the mailchimp api. The response from mailchimp is passed back to the app. Errors currently arent being handled'